Refactor package structure and dependencies
- Reorganize package structure for clarity (e.g., move `InstructorProviderPort`, `UserProviderPort`, and `GymUpdatedEvent` to appropriate packages).
- Replace `@Transactional` in `GymInfoIntegrationTest` with `Awaitility` to ensure proper handling of asynchronous events.
- Add `awaitility` dependency.
- Add `@Profile("!test")` to `DataSeeder` to exclude it from test profile execution.
- Apply minor formatting and import adjustments.
