+++ src/main/java/com/example/boka/gymclass/InstructorProviderPort.java
@@ -1,4 +1,4 @@
-package com.example.boka.gymclass.application;
+package com.example.boka.gymclass;
 
 import java.util.Map;
 import java.util.Set;
