+++ src/main/java/com/example/boka/user/infrastructure/InstructorProviderAdapter.java
@@ -1,6 +1,6 @@
 package com.example.boka.user.infrastructure;
 
-import com.example.boka.gymclass.application.InstructorProviderPort;
+import com.example.boka.gymclass.InstructorProviderPort;
 import com.example.boka.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
